+++ module/data/hq-trait.mjs
@@ -1,0 +1,12 @@
+export default class MarvelMultiverseHqTrait extends foundry.abstract.TypeDataModel {
+  static defineSchema() {
+    const fields = foundry.data.fields;
+    const schema = {};
+
+    schema.description = new fields.StringField({ required: true, blank: true });
+    schema.downtimeActivity = new fields.StringField({ required: true, blank: true });
+    schema.maxCount = new fields.NumberField({ required: true, nullable: false, integer: true, initial: 0, min: 0 });
+
+    return schema;
+  }
+}